BackendMobileAPP/requirements.txt
fastapi~=0.115.12
uvicorn
python-jose~=3.4.0
cryptography
passlib[bcrypt]~=1.7.4
python-multipart
asyncpg
python-dotenv~=1.1.0
SQLAlchemy~=2.0.40
pydantic~=2.11.3
DateTime~=5.5
pytest
alembic~=1.15.2
Django~=5.2
requests~=2.32.3
starlette~=0.46.2
openpyxl~=3.1.5
pydantic[email]
schemas~=0.7.1
pydantic-settings~=2.9.1
colorlog~=6.9.0
logger~=1.4
typing~=3.7.4.3
the-new-hotness~=1.3.0
spglib~=2.6.0
oauthlib~=3.2.2
aiosmtplib
email-validator
twilio
zoomus
